adding item.Omittable interface for field-level omission of data in responses, implementation in handler and separating source into individual files for other interface impls
